tauth-pro/requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
backports.entry-points-selectable==1.1.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
docutils==0.17.1
ecdsa==0.17.0
filelock==3.4.0
idna==2.10
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
packaging==20.9
platformdirs==2.4.0
pyasn1==0.4.8
Pygments==2.9.0
pyparsing==2.4.7
python-jose==3.3.0
python-keycloak==0.26.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
urllib3==1.26.6
virtualenv==20.10.0
virtualenv-clone==0.5.7
